e002-01-gan-cluster-metrics/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open_clip_torch==2.24.0
scikit-learn==1.4.2
pyyaml==6.0.1
scipy==1.11.4
pandas>=2.0.0
numpy>=1.24.0
matplotlib
